+++ omni-sdk/packages/react/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @omni-analytics/react
 
+## 0.2.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @omni-analytics/sdk@0.3.0
+
 ## 0.2.0
 
 ### Minor Changes